backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
email-validator==2.1.1
bcrypt==4.1.2

